requirements.txt
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
cattrs==1.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
coverage==6.3.2
dj-database-url==0.5.0
Django==4.0.3
django-ratelimit==3.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
python-decouple==3.6
pytz==2022.1
requests==2.27.1
requests-cache==0.9.3
six==1.16.0
sqlparse==0.4.2
url-normalize==1.4.3
urllib3==1.26.9
